+++ server/data/data.js
@@ -1,237 +1,239 @@
-const labels =  [
-        {
-          "id": "pork",
-          "label": "Pork"
-        },
-        {
-          "id": "seafood",
-          "label": "Seafood"
-        },
-        {
-          "id": "kids",
-          "label": "Kids"
-        },
-        {
-          "id": "chicken",
-          "label": "Chicken"
-        },
-        {
-          "id": "beef",
-          "label": "Beef"
-        },
-        {
-          "id": "vegetarian",
-          "label": "Vegetarian"
-        },
-        {
-          "id": "breakfast",
-          "label": "Breakfast"
-        }
+const label =  [
+  {
+    "id": "pork",
+    "label": "Pork"
+  },
+  {
+    "id": "seafood",
+    "label": "Seafood"
+  },
+  {
+    "id": "kids",
+    "label": "Kids"
+  },
+  {
+    "id": "chicken",
+    "label": "Chicken"
+  },
+  {
+    "id": "beef",
+    "label": "Beef"
+  },
+  {
+    "id": "vegetarian",
+    "label": "Vegetarian"
+  },
+  {
+    "id": "breakfast",
+    "label": "Breakfast"
+  }
 ]
 
-const meals =  [
-    {
-      "id": "meal1",
-      "title": "3 course chicken",
-      "starter": "Lorem Ipsum",
-      "desert": "Cake",
-      "price": 9.99,
-      "labels": [
-        "chicken",
-        "breakfast"
-      ],
-      "img": "https://source.unsplash.com/XaDsH-O2QXs",
-      "drinks": [
-        {
-          "id": "drink-1",
-          "title": "Vine",
-          "price": 4.99
-        },
-        {
-          "id": "drink-2",
-          "title": "Juice",
-          "price": 5.99
-        },
-        {
-          "id": "drink-3",
-          "title": "Beer",
-          "price": 6.99
-        }
-      ]
-    },
-    {
-      "id": "meal2",
-      "title": "3 course Beef",
-      "starter": "Lorem Ipsum",
-      "desert": "Cake",
-      "price": 19.99,
-      "labels": [
-        "beef"
-      ],
-      "img": "https://source.unsplash.com/auIbTAcSH6E",
-      "drinks": [
-        {
-          "id": "drink-1",
-          "title": "Vine",
-          "price": 4.99
-        },
-        {
-          "id": "drink-2",
-          "title": "Juice",
-          "price": 5.99
-        },
-        {
-          "id": "drink-3",
-          "title": "Beer",
-          "price": 6.99
-        }
-      ]
-    },
-    {
-      "id": "meal3",
-      "title": "3 course Vegetarian",
-      "starter": "Lorem Ipsum",
-      "desert": "Cake",
-      "price": 79.99,
-      "labels": [
-        "vegetarian"
-      ],
-      "img": "https://source.unsplash.com/EvoIiaIVRzU",
-      "drinks": [
-        {
-          "id": "drink-1",
-          "title": "Vine",
-          "price": 4.99
-        },
-        {
-          "id": "drink-2",
-          "title": "Juice",
-          "price": 5.99
-        },
-        {
-          "id": "drink-3",
-          "title": "Beer",
-          "price": 6.99
-        }
-      ]
-    },
-    {
-      "id": "meal4",
-      "title": "3 course Seafood",
-      "starter": "Lorem Ipsum",
-      "desert": "Cake",
-      "price": 49.99,
-      "labels": [
-        "seafood"
-      ],
-      "img": "https://source.unsplash.com/awj7sRviVXo",
-      "drinks": [
-        {
-          "id": "drink-1",
-          "title": "Vine",
-          "price": 4.99
-        },
-        {
-          "id": "drink-2",
-          "title": "Juice",
-          "price": 5.99
-        },
-        {
-          "id": "drink-3",
-          "title": "Beer",
-          "price": 6.99
-        }
-      ]
-    },
-    {
-      "id": "meal5",
-      "title": "3 course Pork",
-      "starter": "Lorem Ipsum",
-      "desert": "Cake",
-      "price": 39.99,
-      "labels": [
-        "pork"
-      ],
-      "img": "https://source.unsplash.com/XPvhzVIeETM",
-      "drinks": [
-        {
-          "id": "drink-1",
-          "title": "Vine",
-          "price": 4.99
-        },
-        {
-          "id": "drink-2",
-          "title": "Juice",
-          "price": 5.99
-        },
-        {
-          "id": "drink-3",
-          "title": "Beer",
-          "price": 6.99
-        }
-      ]
-    },
-    {
-      "id": "meal6",
-      "title": "3 course Kids",
-      "starter": "Lorem Ipsum",
-      "desert": "Cake",
-      "price": 29.99,
-      "labels": [
-        "kids",
-        "breakfast"
-      ],
-      "img": "https://source.unsplash.com/PLyJqEJVre0",
-      "drinks": [
-        {
-          "id": "drink-1",
-          "title": "Vine",
-          "price": 4.99
-        },
-        {
-          "id": "drink-2",
-          "title": "Juice",
-          "price": 5.99
-        },
-        {
-          "id": "drink-3",
-          "title": "Beer",
-          "price": 6.99
-        }
-      ]
-    },
-    {
-      "id": "meal7",
-      "title": "3 course Chicken 2",
-      "starter": "Lorem Ipsum",
-      "desert": "Cake",
-      "price": 19.99,
-      "labels": [
-        "chicken"
-      ],
-      "img": "https://source.unsplash.com/N0u8bLrB_-g",
-      "drinks": [
-        {
-          "id": "drink-1",
-          "title": "Vine",
-          "price": 4.99
-        },
-        {
-          "id": "drink-2",
-          "title": "Juice",
-          "price": 5.99
-        },
-        {
-          "id": "drink-3",
-          "title": "Beer",
-          "price": 6.99
-        }
-      ]
-    }
-  ]
-
-
-  module.exports = {
-    labels,
-    meals
+const meal =  [
+  {
+    "id": "meal1",
+    "title": "3 course chicken",
+    "starter": "Lorem Ipsum",
+    "desert": "Cake",
+    "price": 9.99,
+    "labels": [
+      "chicken",
+      "breakfast"
+    ],
+    "img": "https://static.independent.co.uk/2022/05/02/11/iStock-1044323290.jpg",
+    "drink_flag" : false,
+    "drinks": [
+      {
+        "id": "drink-1",
+        "title": "Vine",
+        "price": 4.99
+      },
+      {
+        "id": "drink-2",
+        "title": "Juice",
+        "price": 5.99
+      },
+      {
+        "id": "drink-3",
+        "title": "Beer",
+        "price": 6.99
+      }
+    ]
+  },
+  {
+    "id": "meal2",
+    "title": "3 course Beef",
+    "starter": "Lorem Ipsum",
+    "desert": "Cake",
+    "price": 19.99,
+    "labels": [
+      "beef"
+    ],
+    "img": "https://www.australianeggs.org.au/assets/blog-posts/211022-Blog-Post-Header-Image-3-Course-Dinner-Party-Ideas__ScaleWidthWzEyMDBd.jpg",
+    "drinks": [
+      {
+        "id": "drink-1",
+        "title": "Vine",
+        "price": 4.99
+      },
+      {
+        "id": "drink-2",
+        "title": "Juice",
+        "price": 5.99
+      },
+      {
+        "id": "drink-3",
+        "title": "Beer",
+        "price": 6.99
+      }
+    ]
+  },
+  {
+    "id": "meal3",
+    "title": "3 course Vegetarian",
+    "starter": "Lorem Ipsum",
+    "desert": "Cake",
+    "price": 79.99,
+    "labels": [
+      "vegetarian"
+    ],
+    "img": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTU9GIpujRfpC8HWO4TPdlH0OA9SDCc2YxEPsPdh0N_YkIzXJh5AUHA3i0_RnJcBSttzaU&usqp=CAU",
+    "drink_flag" : false,
+    "drinks": [
+      {
+        "id": "drink-1",
+        "title": "Vine",
+        "price": 4.99
+      },
+      {
+        "id": "drink-2",
+        "title": "Juice",
+        "price": 5.99
+      },
+      {
+        "id": "drink-3",
+        "title": "Beer",
+        "price": 6.99
+      }
+    ]
+  },
+  {
+    "id": "meal4",
+    "title": "3 course Seafood",
+    "starter": "Lorem Ipsum",
+    "desert": "Cake",
+    "price": 49.99,
+    "labels": [
+      "seafood"
+    ],
+    "img": "https://www.foodandwine.com/thmb/ClPnka2WSnl5PtrMYOjlmXsXw1k=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/escovitch-fish-FT-RECIPE0920-8a733638c2ba4b72b48737782fa616c2.jpg",
+    "drink_flag" : false,
+    "drinks": [
+      {
+        "id": "drink-1",
+        "title": "Vine",
+        "price": 4.99
+      },
+      {
+        "id": "drink-2",
+        "title": "Juice",
+        "price": 5.99
+      },
+      {
+        "id": "drink-3",
+        "title": "Beer",
+        "price": 6.99
+      }
+    ]
+  },
+  {
+    "id": "meal5",
+    "title": "3 course Pork",
+    "starter": "Lorem Ipsum",
+    "desert": "Cake",
+    "price": 39.99,
+    "labels": [
+      "pork"
+    ],
+    "img": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRX3vJZR5sZ1O6eOqEyyJBEWGbe-rTt2mcv-w&s",
+    "drink_flag" : false,
+    "drinks": [
+      {
+        "id": "drink-1",
+        "title": "Vine",
+        "price": 4.99
+      },
+      {
+        "id": "drink-2",
+        "title": "Juice",
+        "price": 5.99
+      },
+      {
+        "id": "drink-3",
+        "title": "Beer",
+        "price": 6.99
+      }
+    ]
+  },
+  {
+    "id": "meal6",
+    "title": "3 course Kids",
+    "starter": "Lorem Ipsum",
+    "desert": "Cake",
+    "price": 29.99,
+    "labels": [
+      "kids",
+      "breakfast"
+    ],
+    "img": "https://qph.cf2.quoracdn.net/main-qimg-c444274bc11709f89001de706002aaf7",
+    "drink_flag" : false,
+    "drinks": [
+      {
+        "id": "drink-1",
+        "title": "Vine",
+        "price": 4.99
+      },
+      {
+        "id": "drink-2",
+        "title": "Juice",
+        "price": 5.99
+      },
+      {
+        "id": "drink-3",
+        "title": "Beer",
+        "price": 6.99
+      }
+    ]
+  },
+  {
+    "id": "meal7",
+    "title": "3 course Chicken 2",
+    "starter": "Lorem Ipsum",
+    "desert": "Cake",
+    "price": 19.99,
+    "labels": [
+      "chicken"
+    ],
+    "img": "https://pataksusa.com/wp-content/uploads/2023/09/3-step-butter-chicken.jpg",
+    "drink_flag" : false,
+    "drinks": [
+      {
+        "id": "drink-1",
+        "title": "Vine",
+        "price": 4.99
+      },
+      {
+        "id": "drink-2",
+        "title": "Juice",
+        "price": 5.99
+      },
+      {
+        "id": "drink-3",
+        "title": "Beer",
+        "price": 6.99
+      }
+    ]
   }
+]
+
+module.exports = {meal, label};
